% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Argentina_taxes_df}
\alias{Argentina_taxes_df}
\title{Monthly Import Taxes of Argentina}
\format{
A data frame with 10 observations and 1 variable:
\describe{
  \item{taxes}{Numeric vector indicating monthly import tax revenue}
}
}
\source{
Data taken from the robustarima package version 0.2.7
}
\usage{
data(Argentina_taxes_df)
}
\description{
This dataset, Argentina_taxes_df, is a data frame containing monthly import tax revenue
in Argentina for the period from January to October 1992.
}
\details{
The dataset name has been kept as 'Argentina_taxes_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ArgentinAPI package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
