% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{detention_centers_tbl_df}
\alias{detention_centers_tbl_df}
\title{Argentina Detention Centers}
\format{
A tibble with 762 observations and 7 variables:
\describe{
  \item{ID}{Numeric ID of the site}
  \item{lugar_emplazamiento_propiedad}{Description of the location and ownership}
  \item{denominacion}{Designation or known name of the detention center}
  \item{espacio_de_memoria}{Logical indicator of whether the site is a memory space}
  \item{ubicacion}{Textual description of the site's geographic location}
  \item{lon}{Longitude}
  \item{lat}{Latitude}
}
}
\source{
Based on data from the presentes package (version 0.1.0).
        Available at: \url{https://www.argentina.gob.ar/sites/default/files/6._anexo_v_listado_de_ccd-investigacion_ruvte-ilid.pdf}
}
\usage{
data(detention_centers_tbl_df)
}
\description{
This dataset, detention_centers_tbl_df, is a tibble containing detailed information
about clandestine detention centers used during Argentina's last dictatorship.
The dataset includes site identifiers, locations, designations, and coordinates.
}
\details{
The dataset name has been kept as 'detention_centers_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ArgentinAPI package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
