% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{STRs_argentina_list}
\alias{STRs_argentina_list}
\title{STRs Allelic Frequencies from Argentina}
\format{
A list with 24 elements, each corresponding to a genetic STR marker:
\describe{
  \item{D8S1179, D21S11, D7S820, ... SE33}{Each element is a named numeric vector of allele frequencies for the respective STR marker.}
}
}
\source{
Data taken from the fbnet package version 1.0.3
}
\usage{
data(STRs_argentina_list)
}
\description{
This dataset, STRs_argentina_list, is a list containing allelic frequencies for 24 Short Tandem Repeat (STR) genetic markers
from populations in Argentina. Each list element corresponds to an STR marker, and contains named numeric vectors with
allele frequencies. The dataset provides valuable reference data for forensic and population genetics studies.
}
\details{
The dataset name has been kept as 'STRs_argentina_list' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ArgentinAPI package and assists users in identifying its specific characteristics.
The suffix 'list' indicates that the dataset is a list object. The original content has not been modified
in any way.
}
