% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argentinapi-package.R
\docType{package}
\name{ArgentinAPI}
\alias{ArgentinAPI}
\alias{ArgentinAPI-package}
\title{ArgentinAPI: Access Argentine Economic, Social, and Geopolitical Data via RESTful APIs and Curated Datasets}
\description{
{
This package provides functions to access data from the 'ArgentinaDatos API' and the 'REST Countries API' related to Argentina's exchange rates, inflation, political figures, holidays, and general country-level indicators, additionally, the package includes curated datasets related to Argentina, covering topics such as economic indicators, biodiversity, agriculture, human rights, genetic data, and consumer prices.
}
}
\details{
ArgentinAPI: Access Argentine Economic, Social, and Geopolitical Data via RESTful APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Argentine Economic, Social, and Geopolitical Data via RESTful APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/argentinapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
